Added a new test to test RNN with a custom layer. Made changes for style conformance.
